+++ tests/testthat/test-ec.util.R
@@ -5,11 +5,11 @@ test_that("serie from ec.util with cartesian3D", {
   tmp <- st_as_sf(data.frame(x=c(-70,-70,-70), y=c(45, 46, 47), z=c(1,2,3)), 
                   coords= c('x','y','z'), crs= st_crs(4326))
   p <- ec.init(load= c('3D'),
-               series= ec.util(df= tmp, 
-                               coordinateSystem= 'cartesian3D', type= 'scatter3D')
-               #,tooltip= list(formatter= '{b}')
+               series= ec.util(df= tmp, cs= 'cartesian3D')
+               ,tooltip= list(formatter= '{b}')
   )
-  expect_equal(p$x$opts$series[[1]]$data[[2]][[2]], 46)
+  expect_s3_class(p$x$opts$series[[1]]$data[[2]]$value, 'sfg')
+  expect_equal(as.numeric(p$x$opts$series[[1]]$data[[2]]$value), c(-70,46,2))
   expect_type( p$x$opts$xAxis3D[[1]],'list')
 })
 
@@ -20,28 +20,29 @@ test_that("shapefiles with multi-polygons", {
   p <- ec.init(load= c('leaflet', 'custom'),  # load custom for polygons
                js= ec.util(cmd= 'sf.bbox', bbox= st_bbox(nc$geometry)),
                series= ec.util(cmd= 'sf.series', df= nc, nid= 'NAME', itemStyle= list(opacity= 0.3)),
-               tooltip= list(show= TRUE, formatter= '{a}')
+               tooltip= list(formatter= '{a}')
   )
   expect_true(p$x$opts$leaflet$roam)
   expect_equal(p$x$opts$series[[108]]$name, 'Brunswick')
   expect_equal(p$x$opts$series[[108]]$itemStyle$opacity, 0.3)
 })
 
-test_that("shapefile from ZIP", {
+test_that("shapefile LINES from ZIP", {
   if (interactive()) {  # creates a subfolder 'railways'
     library(sf)
     fname <- ec.util(cmd= 'sf.unzip', 
                      url= 'https://mapcruzin.com/sierra-leone-shapefiles/railways.zip')
     nc <- as.data.frame(st_read(fname))
     p <- ec.init(load= 'leaflet',
-                 js= ec.util(cmd= 'sf.bbox', bbox= st_bbox(nc$geometry)), 
-                 series= ec.util(df= nc, nid= 'osm_id', verbose=TRUE,
-                                 lineStyle= list(width= 3, color= 'red')),
-                 tooltip= list(formatter= '{a}'), animation= FALSE,
-                 leaflet= list(tiles= list(list(
-                   urlTemplate= 'https://stamen-tiles-{s}.a.ssl.fastly.net/terrain/{z}/{x}/{y}{r}.{ext}',
-                   options= list(attribution= 'Map tiles by <a href="http://stamen.com">Stamen Design</a>, <a href="http://creativecommons.org/licenses/by/3.0">CC BY 3.0</a>',
-                                 subdomains= 'abcd',	maxZoom= 18,	ext= 'png')))) 
+       js= ec.util(cmd= 'sf.bbox', bbox= st_bbox(nc$geometry)), 
+       series= ec.util(df= nc, nid= 'osm_id', verbose=TRUE,
+                       lineStyle= list(width= 3, color= 'red')),
+       tooltip= list(formatter= '{a}'), animation= FALSE,
+       leaflet= list( roam= TRUE,
+         tiles= list(list(
+           urlTemplate= 'https://stamen-tiles-{s}.a.ssl.fastly.net/terrain/{z}/{x}/{y}{r}.{ext}',
+           options= list(attribution= 'Map tiles by <a href="http://stamen.com">Stamen Design</a>, <a href="http://creativecommons.org/licenses/by/3.0">CC BY 3.0</a>',
+                         subdomains= 'abcd', maxZoom= 18, ext= 'png')))) 
     )
     expect_equal(p$x$opts$leaflet$tiles[[1]]$options$subdomains, 'abcd')
     expect_equal(p$x$opts$series[[6]]$name, '207557821')
@@ -51,15 +52,39 @@ test_that("shapefile from ZIP", {
   else expect_equal(1,1)
 })
 
-test_that("tabset", {
-  p1 <- cars |> ec.init(width= 300, height= 300, grid= list(top= 20))
-  p2 <- mtcars |> ec.init(width= 300, height= 300)
-  r <- htmltools::browsable(
-    ec.util(cmd='tabset', cars=p1, mtcars=p2)
-  )
+test_that("shapefile POINTS from ZIP", {
+  if (interactive()) {  # creates a subfolder 'points'
+    library(sf)
+    fn <- ec.util(cmd= 'sf.unzip', 
+                  url= 'https://mapcruzin.com/sierra-leone-shapefiles/points.zip')
+    nc <- as.data.frame(st_read(fn)) |> head(10)
+    p <- ec.init(load= c('leaflet'),
+       js= ec.util(cmd= 'sf.bbox', bbox= st_bbox(nc$geometry)), 
+       series= ec.util(df= nc, name= 'spots', itemStyle= list(color= 'red'), verbose=TRUE),
+       tooltip= list(valueFormatter= ec.clmn('json')), legend= list(show= TRUE)
+    )
+    expect_s3_class(p$x$opts$series[[1]]$data[[2]]$value, 'sfg')
+    expect_equal(round(as.numeric(p$x$opts$series[[1]]$data[[2]]$value),1), c(-13.3,8.5))
+    expect_true( p$x$opts$leaflet$roam)
+  }
+  else expect_equal(1,1)
+})
+
+test_that("ec.util layout", {
+  p <- lapply(list('dark','macarons','gray','jazz','dark-mushroom'),
+              function(x) cars |> ec.init() |> ec.theme(x) ) |>
+    ec.util(cmd='layout', cols= 2, title= 'my layout')
+  expect_equal(p$children[[3]]$children[[1]]$children[[2]]$children[[1]]$x$theme, 'macarons')
+})
+
+test_that("tabset with pairs", {
+  p1 <- cars |> ec.init(grid= list(top= 20))
+  p2 <- mtcars |> ec.init()
+  r <- ec.util(cmd='tabset', cars=p1, mtcars=p2)
   expect_equal(r[[2]]$children[[5]]$children[[1]]$children[[1]][[1]]$x$opts$dataset[[1]]$source[[1]], c("speed", "dist"))
   expect_equal(r[[2]]$children[[5]]$children[[1]]$name, "section")
   expect_equal(r[[2]]$children[[2]]$children[[1]], "cars")
+  expect_s3_class(r[[2]]$children[[5]]$children[[2]]$children[[1]][[1]], 'echarty')
 })
 
 test_that("tabset with pipe", {
@@ -69,7 +94,7 @@ test_that("tabset with pipe", {
       x |> ec.init(ctype= 'scatter', title= list(text= unique(x$Species)))
     }) |> ec.util(cmd='tabset')
   )
-  expect_equal(r[[2]]$children[[7]]$children[[2]]$children[[1]][[1]]$width, 300)
+  expect_equal(r[[2]]$children[[7]]$children[[2]]$children[[1]][[1]]$width, "100%")
   expect_equal(r[[2]]$children[[6]]$children[[1]], "chart3")
 })
 
@@ -119,6 +144,24 @@ test_that("morph", {
   expect_equal(p$x$on[[1]]$event, 'mouseover')
 })
 
+test_that("ec.util morph", {
+  setd <- function(type) {
+    mtcars |> group_by(cyl) |> ec.init(ctype= type) |> ec.upd({
+      title <- list(subtext='mouseover points to morph')
+      xAxis <- list(scale=TRUE)
+      series <- lapply(series, function(ss) {
+        ss$groupId <- ss$name
+        ss$universalTransition <- list(enabled=TRUE)
+        ss })
+    })
+  }
+  oscatter <- setd('scatter')
+  obar <- setd('bar')
+  p <- ec.util(cmd='morph', oscatter, obar)
+  expect_equal(p$x$opts$morph[[2]]$series[[3]]$type, 'bar')
+  expect_true (p$x$opts$morph[[2]]$series[[3]]$universalTransition$enabled)
+})
+
 test_that("fullscreen", {
   tbox <- list(right='20%', feature= ec.util(cmd='fullscreen'))
   #p <- cars |> ec.init(toolbox= tbox)
@@ -152,10 +195,9 @@ test_that("level", {
   expect_equal(p, c(1,2,1))
 })
 
-test_that("labelsInside", {
+test_that("labelsInside and doType", {
   p <- ec.init(
     xAxis= list(data= list(1,2,3,4,5,6,7)),
-    yAxis= list(type= 'value'),
     series= list(
       list(name= 'long text, 20 chars', type='line',
            data= c(110, 132, 101, 134, 90, 230, 210),
@@ -168,4 +210,5 @@ test_that("labelsInside", {
     )
   )
   expect_match(p$x$opts$series[[2]]$labelLayout, "get_e_charts(cid)", fixed=TRUE)
+  expect_equal(p$x$opts$xAxis$type, 'category')
 })